Fixes after testing
In task No. 5, the error of the difference between the background color and the span name was corrected.
In task No. 6, the createBoxes function is supplemented in such a way that elements are added to the DOM in one operation. Also, the name of the deleteEl function has been changed to the correct deleteEl